Split out client and server gulp files
