+++ gulpfile.js
@@ -2,26 +2,19 @@
 
 let
 	_ = require('lodash'),
+	fs = require('fs'),
 	path = require('path'),
 	chalk = require('chalk'),
 	colors = require('colors'),
 	runSequence = require('run-sequence'),
 	gulp = require('gulp'),
 	plugins = require('gulp-load-plugins')(),
-	glob = require('glob'),
-	argv = require('yargs').argv,
-	del = null,
-	webpack = null,
 
-	pkg = require('./package.json'),
 	assets = require(path.resolve('./config/assets.js'));
 
-// These two dependencies are only needed for client builds
-// If the mean2-starter is installed with yeoman, these dependencies may not exist in package.json
-if (assets.client) {
-	webpack = require('webpack');
-	del = require('del');
-}
+// Initialize gulp-submodule
+require('gulp-submodule')(gulp);
+
 
 // Patch chalk to use colors
 chalk.enabled = true;
@@ -30,314 +23,53 @@ colors.enabled = true;
 
 /**
  * --------------------------
- * Watch/Reload Tasks
+ * Submodule tasks
  * --------------------------
  */
 
+let hasClient = false, hasServer = false;
+
+if (fs.existsSync('./src/server/gulpfile.js')) {
+	hasServer = true;
+	gulp.submodule('server', { filepath: './src/server/gulpfile.js' });
+}
+
+if (fs.existsSync('./src/client/gulpfile.js')) {
+	hasClient = true;
+	gulp.submodule('client', { filepath: './src/client/gulpfile.js' });
+}
+
 /**
- * Nodemon helper function. This function will start nodemon
- * using the provided configuration. It also registers handlers
- * for the crash and restart events to log the events.
+ * @callback GulpCallback
+ * @param {Object} err
+ */
+
+/**
+ * Run a sequence of gulp commands. Any server or client tasks will be ignored
+ * if the respective sections don't exist.
  *
- * @param nodemonConfig The config to pass to nodemon
- * @returns {*}
+ * @param {...(string|string[]|function)} tasks - A gulp task to run, or an array of tasks to run in parallel.
+ *   The last argument must be a callback to invoke when all the tasks have completed.
  */
-function nodemon(nodemonConfig) {
-	let config = require('./src/server/config');
-	let nodeArgs = ['--debug=' + config.devPorts.debug, '--inspect'];
-	if (plugins.util.env.debugBrk) {
-		nodeArgs.push('--debug-brk');
+function run(tasks) {
+	function isValidPath(path) {
+		// Recursively iterate into arrays of parallel tasks
+		if (_.isArray(path)) {
+			let filteredArray = _.remove(path, isValidPath);
+			return filteredArray.length === 0;
+		}
+		if (!hasServer && path.startsWith('server')) {
+			return true;
+		}
+		if (!hasClient && path.startsWith('client')) {
+			return true;
+		}
+		return false;
 	}
-	nodemonConfig.nodeArgs = nodeArgs;
 
-	let stream = plugins.nodemon(nodemonConfig);
-	stream
-		.on('restart', () => {
-			plugins.util.log('Restarting server...');
-		})
-		.on('crash', () => {
-			plugins.util.log('Application crashed, restarting...');
-			stream.emit('restart', 1);
-		});
-
-	return stream;
+	let filteredPaths = _.remove(_.initial(arguments), isValidPath);
+	runSequence(filteredPaths, _.last(arguments));
 }
-
-gulp.task('watch-server', () => {
-	return nodemon({
-		script: 'src/server/server.js',
-		ext: 'js json html',
-		watch: _.union(
-			assets.build,
-			assets.server.views,
-			assets.server.allJS,
-			assets.server.config),
-		tasks: [ 'build-server' ]
-	});
-});
-
-
-gulp.task('watch-client', () => {
-	if (!assets.client) return;
-
-	var config = require('./src/server/config');
-
-	// Start livereload
-	plugins.livereload.listen(config.devPorts.liveReload);
-
-	/*
-	 * Add watch rules to trigger livereload
-	 */
-
-	// On changes to compiled stuff, recompile
-	gulp.watch(assets.client.app.src.sass, ['build-client-style']);
-
-	// In dev mode, we just want to re-lint ts code
-	gulp.watch(assets.client.app.src.ts, ['lint-client-code'])
-		.on('change', (d) => { setTimeout(() => { plugins.livereload.changed(d); }, 1000); });
-
-	// When generated css changes, let livereload handle the changes
-	gulp.watch(assets.client.app.dist.development.css).on('change', plugins.livereload.changed);
-
-	// When views or content change, force livereload to reload the whole page (we had issues with changes getting missed)
-	gulp.watch(assets.client.app.views).on('change', () => { setTimeout(plugins.livereload.reload, 1000); });
-	gulp.watch(assets.client.app.content).on('change', () => { setTimeout(plugins.livereload.reload, 1000); });
-});
-
-
-
-/**
- * --------------------------
- * Server Build Tasks
- * --------------------------
- */
-gulp.task('build-server', () => {
-	return gulp.src(_.union(assets.server.allJS, assets.tests.server, assets.build))
-		// ESLint
-		.pipe(plugins.eslint('./config/build/eslint.conf.json'))
-		.pipe(plugins.eslint.format())
-		.pipe(plugins.eslint.failAfterError());
-});
-
-
-/**
- * --------------------------
- * Client Build Tasks
- * --------------------------
- */
-
-gulp.task('build-client', (done) => {
-	if (!assets.client) return done();
-
-	runSequence('clean-client', ['build-client-code', 'build-client-style'], done);
-});
-
-gulp.task('clean-client', () => {
-	if (!assets.client) return;
-
-	return del([ 'public/**/*' ]);
-});
-
-gulp.task('build-client-code', ['lint-client-code'], (done) => {
-	if (!assets.client) return done();
-
-	let webpackConfig = require(path.resolve('./config/build/webpack.conf.js'));
-
-	webpack(webpackConfig('build'), (err, stats) => {
-
-		// Fail if there were errors
-		if(err) throw new plugins.util.PluginError('webpack', err);
-
-		// log the stats from webpack
-		plugins.util.log('[webpack]', stats.toString({
-			colors: true, chunks: false
-		}));
-
-		done();
-	});
-});
-
-gulp.task('lint-client-code', () => {
-	if (!assets.client) return;
-
-	// Grab the tslint config
-	var config = require(path.resolve('./config/build/tslint.conf.js'));
-	config.formatter = 'prose';
-
-	return gulp.src(assets.client.app.src.ts)
-		// Lint the Typescript
-		.pipe(plugins.tslint(config))
-		.pipe(plugins.tslint.report({
-			summarizeFailureOutput: true,
-			emitError: true
-		}));
-
-});
-
-gulp.task('build-client-style', [ 'clean-client-style' ], () => {
-	if (!assets.client) return;
-
-	// Generate a list of the sources in a deterministic manner
-	let sourceArr = [];
-	assets.client.app.src.sass.forEach((f) => {
-		sourceArr = sourceArr.concat(glob.sync(f).sort());
-	});
-
-	return gulp.src(sourceArr)
-
-		// Lint the Sass
-		.pipe(plugins.sassLint({
-			formatter: 'stylish',
-			rules: require(path.resolve('./config/build/sasslint.conf.js'))
-		}))
-		.pipe(plugins.sassLint.format())
-		.pipe(plugins.sassLint.failOnError())
-
-		// Compile and concat the sass (w/sourcemaps)
-		.pipe(plugins.sourcemaps.init())
-			.pipe(plugins.sass())
-			.pipe(plugins.concat('application.css'))
-			.pipe(plugins.insert.prepend(assets.bannerString))
-		.pipe(plugins.sourcemaps.write('.', { sourceRoot: null }))
-		.pipe(gulp.dest('public/dev'))
-
-		// Clean the CSS
-		.pipe(plugins.filter('public/dev/application.css'))
-		.pipe(plugins.cleanCss())
-		.pipe(plugins.insert.prepend(assets.bannerString))
-		.pipe(plugins.hash({
-			template: 'application.<%= hash %>.min.css',
-			hashLength: 16
-		}))
-		.pipe(gulp.dest('public'));
-});
-
-gulp.task('clean-client-style', () => {
-	if (!assets.client) return;
-
-	return del([
-		'public/application*.css',
-		'public/dev/application.css', 'public/dev/application.css.map'
-	]);
-});
-
-
-/**
- * --------------------------
- * Testing Tasks
- * --------------------------
- */
-
-gulp.task('env:test', () => {
-	// Set the environment to test
-	process.env.NODE_ENV = 'test';
-});
-
-
-gulp.task('test-server', ['env:test'], () => {
-	// Gather some args for custom testing
-	let args = [];
-
-	// --bail will cause mocha to stop on first error
-	if(argv.bail) { args.push('--bail'); }
-
-	// --filter will filter the test files using the glob pattern
-	if(null != argv.filter) { args.push(`--filter='${argv.filter}'`); }
-
-	// Run mocha tests with nodemon
-	return nodemon({
-		script: './config/build/test-server.js',
-		ext: 'js json',
-		env: { 'NODE_ENV': 'test' },
-		args: args,
-		watch: _.union(
-			assets.tests.server,
-			assets.server.allJS,
-			assets.server.config,
-			assets.build)
-	});
-});
-
-function runKarmaTest(callback) {
-	const spawn = require('child_process').spawn;
-	const karma = spawn('node', ['./config/build/test-client.js']);
-
-	karma.stdout.pipe(process.stdout);
-	karma.stderr.pipe(process.stderr);
-
-	karma.on('close', callback);
-}
-
-gulp.task('test-client', ['env:test'], () => {
-
-	const clientFilesToWatch = _.union(
-			assets.tests.client,
-			assets.client.app.src.ts,
-			assets.client.app.src.sass,
-			assets.client.app.views,
-			assets.client.app.content,
-			assets.build
-	);
-
-	const minTimeBetweenTestsCalls = 5000;
-
-	const runKarmaTestWithDebounce = _.throttle(_.partial(runKarmaTest, _.noop), minTimeBetweenTestsCalls);
-
-	gulp.watch(clientFilesToWatch, runKarmaTestWithDebounce);
-	runKarmaTestWithDebounce();
-
-});
-
-gulp.task('test-client-ci', ['env:test'], runKarmaTest);
-
-gulp.task('coverage-init', () => {
-	// Covering all server code minus routes
-	return gulp.src([
-			'src/server/**/*.js',
-			'!src/server/**/routes/**/*.js'
-		])
-		// Covering files
-		.pipe(plugins.istanbul({
-			includeUntested: true
-		}))
-		// Force `require` to return covered files
-		.pipe(plugins.istanbul.hookRequire());
-});
-
-gulp.task('test-server-ci', [ 'env:test', 'coverage-init' ], (done) => {
-	// Run mocha tests with coverage and without nodemon
-
-	// Open mongoose connections
-	let mongoose = require('./src/server/lib/mongoose.js');
-
-	let error = null;
-	mongoose.connect().then(() => {
-		let sources = assets.tests.server;
-
-		gulp.src(sources)
-			.pipe(plugins.mocha({
-				reporter: 'mochawesome',
-				reporterOptions: {
-					reportDir: 'reports/tests',
-					reportName: 'index',
-					reportTitle: pkg.name + ' Server Test Results',
-					inlineAssets: true
-				}
-			}))
-			.pipe(plugins.istanbul.writeReports({
-				dir: './reports/coverage',
-				reporters: ['html']
-			}))
-			.on('error', (err) => {
-				error = err;
-			})
-			.on('end', () => {
-				mongoose.disconnect().then(() => {
-					done(error);
-				}).done();
-			});
-	}).done();
-});
 
 
 /**
@@ -351,9 +83,9 @@ gulp.task('test-server-ci', [ 'env:test', 'coverage-init' ], (done) => {
  * dev - main development task to run and watch server
  */
 gulp.task('dev', (done) => {
-	runSequence(
-		[ 'build-server', 'build-client-style', 'lint-client-code' ],
-		[ 'watch-server', 'watch-client' ],
+	run(
+		[ 'server:build', 'client:build-style', 'client:lint-code' ],
+		[ 'server:watch', 'client:watch' ],
 		done);
 });
 
@@ -361,19 +93,19 @@ gulp.task('dev', (done) => {
 /**
  * build - Builds production assets into ./public dir
  */
-gulp.task('build', [ 'build-client', 'build-server' ]);
+gulp.task('build', (done) => { run([ 'server:build', 'client:build' ], done); });
 
 
 /**
  * test - Run tests in dev mode with nodemon
  */
-gulp.task('test', (done) => { runSequence([ 'test-server', 'test-client' ], done); });
+gulp.task('test', (done) => { run([ 'server:test', 'client:test' ], done); });
 
 
 /**
  * Run tests in CI mode with coverage and no nodemon
  */
-gulp.task('test-ci', (done) => { runSequence([ 'test-server-ci', 'test-client-ci' ], done); });
+gulp.task('test-ci', (done) => { run([ 'server:test-ci', 'client:test-ci' ], done); });
 
 
 /**
