+++ package.json
@@ -1,153 +1,139 @@
 {
-	"name": "mean2",
-	"description": "MEAN starter with MongoDB, Express, Angular 2, and Node.",
-	"version": "0.1.0-SNAPSHOT",
-
-	"private": false,
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/Asymmetrik/mean2.git"
-	},
-
-	"engines": {
-		"node": "6.7.0",
-		"npm": "3.10.9"
-	},
-
-	"scripts": {
-		"start": "node /src/server/server.js",
-		"debug": "gulp dev",
-		"test": "gulp test"
-	},
-
-	"dependencies": {
-		"async": "2.1",
-		"body-parser": "1.15",
-		"bunyan": "1.8",
-		"bunyan-logstash": "0.3",
-		"chalk": "1.1",
-		"colors": "1.1",
-		"compression": "1.6",
-		"connect-flash": "0.1",
-		"connect-mongo": "1.3",
-		"cookie-parser": "1.4",
-		"csv-stringify": "1.0",
-		"express": "4.14",
-		"express-handlebars": "3.0",
-		"express-session": "1.14",
-		"glob": "7.1",
-		"helmet": "3.1",
-		"JSONPath": "0.11",
-		"lodash": "4.17",
-		"method-override": "2.3",
-		"moment": "2.17",
-		"mongoose": "4.6",
-		"mongoose-schema-extend": "0.2.0",
-		"mongoose-unique-validator": "1.0",
-		"morgan": "1.7",
-		"multer": "1.2",
-		"multipipe": "1.0",
-		"nodemailer": "2.6",
-		"node-uuid": "1.4.7",
-		"platform": "1.3.3",
-		"passport": "0.3",
-		"passport-local": "1.0",
-		"path": "0.12",
-		"q": "1.4",
-		"serve-favicon": "2.3",
-		"socket.io": "1.5",
-		"through2": "2.0",
-		"validator": "6.1",
-
-		"kafka-node": "0.5"
-	},
-	"devDependencies": {
-		"@angular/common": "2.1.1",
-		"@angular/compiler": "2.1.1",
-		"@angular/core": "2.1.1",
-		"@angular/forms": "2.1.1",
-		"@angular/http": "2.1.1",
-		"@angular/platform-browser": "2.1.1",
-		"@angular/platform-browser-dynamic": "2.1.1",
-		"@angular/upgrade": "2.1.1",
-		"@angular/router": "3.1.1",
-
-		"core-js": "2.4",
-		"reflect-metadata": "0.1",
-		"rxjs": "5.0.0-rc.4",
-		"zone.js": "0.7",
-
-		"angular2-modal": "2.0.1",
-		"angular2-toaster": "1.0.2",
-		"bootstrap": "3",
-		"font-awesome": "4.7",
-		"ng2-bootstrap": "1.1",
-		"ng2-select": "1.1.2",
-
-		"socket.io-client": "1.4",
-		"@types/socket.io-client": "1.4",
-
-		"@types/core-js": "0.9",
-		"@types/node": "6.0",
-		"@types/lodash": "4.14",
-
-		"webpack": "2.1.0-beta.27",
-		"angular2-template-loader": "0.6",
-		"css-loader": "0.26",
-		"file-loader": "0.9",
-		"html-loader": "0.4",
-		"resolve-url-loader": "1.6",
-		"sass-loader": "4.0",
-		"style-loader": "0.13",
-		"ts-loader": "1.2",
-		"url-loader": "0.5",
-		"webpack-dev-middleware": "1.8",
-		"webpack-hot-middleware": "2.13",
-
-		"typescript": "2.0",
-		"ts-helpers": "1.1",
-		"tslint": "3.15",
-		"tslint-stylish": "2.1.0-beta",
-
-		"mocha": "3.1",
-		"mochawesome": "1.5",
-		"should": "11.1",
-
-		"karma": "1.3.0",
-		"karma-jasmine": "1.1.0",
-		"karma-mocha-reporter": "2.2.1",
-		"karma-phantomjs-launcher": "1.0.2",
-		"karma-sourcemap-loader": "0.3.7",
-		"karma-webpack": "1.8.0",
-		"jasmine-core": "2.5.2",
-		"@types/jasmine": "2.5.38",
-
-		"gulp": "3.9",
-		"del": "2.2",
-		"gulp-clean-css": "2.0",
-		"gulp-concat": "2.6",
-		"gulp-coverage": "0.3",
-		"gulp-eslint": "3.0",
-		"gulp-filter": "4.0",
-		"gulp-hash": "3.1",
-		"gulp-insert": "0.5",
-		"gulp-istanbul": "1.1",
-		"gulp-karma": "0.0.5",
-		"gulp-livereload": "3.8",
-		"gulp-load-plugins": "1.4",
-		"gulp-mocha": "3.0",
-		"gulp-nodemon": "2.2",
-		"gulp-rename": "1.2",
-		"gulp-sass": "2.3",
-		"gulp-sass-lint": "1.3",
-		"gulp-sort": "2.0",
-		"gulp-sourcemaps": "2.2",
-		"gulp-tslint": "6.1",
-		"gulp-util": "3.0",
-		"gulp-watch": "4.3",
-		"run-sequence": "1.2",
-		"node-sass": "3.13",
-		"yargs": "6.4"
-
-	}
+  "name": "mean2",
+  "description": "MEAN starter with MongoDB, Express, Angular 2, and Node.",
+  "version": "0.1.0-SNAPSHOT",
+  "private": false,
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/Asymmetrik/mean2.git"
+  },
+  "engines": {
+    "node": "6.7.0",
+    "npm": "3.10.9"
+  },
+  "scripts": {
+    "start": "node /src/server/server.js",
+    "debug": "gulp dev",
+    "test": "gulp test"
+  },
+  "dependencies": {
+    "JSONPath": "0.11",
+    "async": "2.1",
+    "body-parser": "1.15",
+    "bunyan": "1.8",
+    "bunyan-logstash": "0.3",
+    "chalk": "1.1",
+    "colors": "1.1",
+    "compression": "1.6",
+    "connect-flash": "0.1",
+    "connect-mongo": "1.3",
+    "cookie-parser": "1.4",
+    "csv-stringify": "1.0",
+    "express": "4.14",
+    "express-handlebars": "3.0",
+    "express-session": "1.14",
+    "glob": "7.1",
+    "gulp-submodule": "^0.1.3",
+    "helmet": "3.1",
+    "kafka-node": "0.5",
+    "lodash": "4.17",
+    "method-override": "2.3",
+    "moment": "2.17",
+    "mongoose": "4.6",
+    "mongoose-schema-extend": "0.2.0",
+    "mongoose-unique-validator": "1.0",
+    "morgan": "1.7",
+    "multer": "1.2",
+    "multipipe": "1.0",
+    "node-uuid": "1.4.7",
+    "nodemailer": "2.6",
+    "passport": "0.3",
+    "passport-local": "1.0",
+    "path": "0.12",
+    "platform": "1.3.3",
+    "q": "1.4",
+    "serve-favicon": "2.3",
+    "socket.io": "1.5",
+    "through2": "2.0",
+    "validator": "6.1"
+  },
+  "devDependencies": {
+    "@angular/common": "2.1.1",
+    "@angular/compiler": "2.1.1",
+    "@angular/core": "2.1.1",
+    "@angular/forms": "2.1.1",
+    "@angular/http": "2.1.1",
+    "@angular/platform-browser": "2.1.1",
+    "@angular/platform-browser-dynamic": "2.1.1",
+    "@angular/upgrade": "2.1.1",
+    "@angular/router": "3.1.1",
+    "core-js": "2.4",
+    "reflect-metadata": "0.1",
+    "rxjs": "5.0.0-rc.4",
+    "zone.js": "0.7",
+    "angular2-modal": "2.0.1",
+    "angular2-toaster": "1.0.2",
+    "bootstrap": "3",
+    "font-awesome": "4.7",
+    "ng2-bootstrap": "1.1",
+    "ng2-select": "1.1.2",
+    "socket.io-client": "1.4",
+    "@types/socket.io-client": "1.4",
+    "@types/core-js": "0.9",
+    "@types/node": "6.0",
+    "@types/lodash": "4.14",
+    "webpack": "2.1.0-beta.27",
+    "angular2-template-loader": "0.6",
+    "css-loader": "0.26",
+    "file-loader": "0.9",
+    "html-loader": "0.4",
+    "resolve-url-loader": "1.6",
+    "sass-loader": "4.0",
+    "style-loader": "0.13",
+    "ts-loader": "1.2",
+    "url-loader": "0.5",
+    "webpack-dev-middleware": "1.8",
+    "webpack-hot-middleware": "2.13",
+    "typescript": "2.0",
+    "ts-helpers": "1.1",
+    "tslint": "3.15",
+    "tslint-stylish": "2.1.0-beta",
+    "mocha": "3.1",
+    "mochawesome": "1.5",
+    "should": "11.1",
+    "karma": "1.3.0",
+    "karma-jasmine": "1.1.0",
+    "karma-mocha-reporter": "2.2.1",
+    "karma-phantomjs-launcher": "1.0.2",
+    "karma-sourcemap-loader": "0.3.7",
+    "karma-webpack": "1.8.0",
+    "jasmine-core": "2.5.2",
+    "@types/jasmine": "2.5.38",
+    "gulp": "3.9",
+    "del": "2.2",
+    "gulp-clean-css": "2.0",
+    "gulp-concat": "2.6",
+    "gulp-coverage": "0.3",
+    "gulp-eslint": "3.0",
+    "gulp-filter": "4.0",
+    "gulp-hash": "3.1",
+    "gulp-insert": "0.5",
+    "gulp-istanbul": "1.1",
+    "gulp-karma": "0.0.5",
+    "gulp-livereload": "3.8",
+    "gulp-load-plugins": "1.4",
+    "gulp-mocha": "3.0",
+    "gulp-nodemon": "2.2",
+    "gulp-rename": "1.2",
+    "gulp-sass": "2.3",
+    "gulp-sass-lint": "1.3",
+    "gulp-sort": "2.0",
+    "gulp-sourcemaps": "2.2",
+    "gulp-tslint": "6.1",
+    "gulp-util": "3.0",
+    "gulp-watch": "4.3",
+    "run-sequence": "1.2",
+    "node-sass": "3.13",
+    "yargs": "6.4"
+  }
 }
